pyproject.toml
[project]
name = "inaction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.8.2",
    "xmltodict>=0.13.0",
    "types-xmltodict>=0.13.0.3",
    "types-sqlalchemy>=1.4.53.38",
    "typing-extensions>=4.12.2",
    "loguru>=0.7.2",
    "result",
    "sneklib[snektest]>=0.0.1",
    "aiosqlite",
]

[project.optional-dependencies]
dev = [
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.2",
]

[tool.setuptools.packages.find]
where = ["app"]

[tool.uv.sources]
result = { git = "https://github.com/rustedpy/result" }
aiosqlite = { git = "https://github.com/crpier/aiosqlite.git" }

[[tool.mypy]]
enable_incomplete_feature = ["NewGenericSyntax"]
